GestureSwipeLab
I tried to have 2 box views overlap a label. The point was to have a label be swipeable and trigger one of 2 commands connected to the action.
But because they overlapped, they seemed to conflict with one another.

As a result, I had to split the label, with a color code to indicate which side corresponds with what.